Создает модульность
