+++ js/data.js
@@ -1,0 +1,82 @@
+import {getRandomInteger} from './util.js';
+import {getRandomArrayElement} from './util.js';
+import {createIdGenerator} from './util.js';
+
+const NAMES = [
+  'Аксинья',
+  'Берта',
+  'Вадим',
+  'Генрих',
+  'Джозеф',
+  'Женя',
+  'Жанна',
+  'Зоя',
+  'Игорь',
+  'Ксения',
+];
+
+const DESCRIPTIONS = [
+  'Цвет — все, но черно-белый — еще больше.',
+  'Предназначение концептуальной фотографии в том, чтобы заставить нас думать.',
+  'Мою жизнь формирует неудержимая потребность странствовать и наблюдать, а мой фотоаппарат — это паспорт.',
+  'Вдохновение никогда еще не бывает таким сильным, как во время путешествий.',
+  'Пейзажная фотография переступает все политические и национальные границы.',
+  'Фотография — это кнопка паузы, на которую поставлена жизнь.',
+  'Не каждый верит картинам, но фотографиям верят все.',
+  'Жанр аэрофотографии тесно связан с дрон-фотографией, но так было не всегда.',
+  'Фотография уловила факт из жизни, и этот факт будет жить вечно.',
+  'Главные герои уличной фотографии - люди, события и само течение жизни.',
+  'Одно дело — сфотографировать, как человек выглядит, и совсем другое, отобразить, что он из себя представляет.',
+  'Суть фотографии в том, чтобы найти что-то интересное в обычном месте.',
+  'Черно-белые снимки — это монохромное искусство.',
+  'Новые места, люди и неизведанная культура — мечта любого фотографа.',
+  'Мир полон завораживающих, непостижимых и впечатляющих видов.',
+  'Любой профессиональный фотограф подтвердит, что его работа требует творческой жилки.',
+  'HDR-фотография чаще всего применяется в пейзажной фотосъемке.',
+  'Аэрофотография предполагает использование широкоугольной фотосъемки.',
+  'Поймать сиюминутный момент — задача не из легких, и этот опыт способен преобразить любого фотографа.',
+  'Тема здорового образа жизни пользуется в стоковой фотографии неизменной популярностью.',
+  'Портретная фотография — это, прежде всего, люди.',
+  'Задача фотографа в предметной фотографии — показать что-то абсолютно обычное в совершенно новом свете.',
+  'Автор черно-белой фотографии сознательно ограничивает себя в цвете, чтобы сконцентрироваться на других выразительных способах.',
+  'Трэвел-фотограф отправился в далекие странствия, чтобы удивить и показать нам то, чего мы еще никогда не видели.',
+  'Алекс - один из наших самых популярных фотографов, работает с молодыми парами и вообще много снимает людей.',
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const PHOTO_DESCRIPTION_COUNT = 25;
+
+const createComment = () => ({
+  id: getRandomInteger(1, 10),
+  avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
+  message: getRandomArrayElement(MESSAGES),
+  name: getRandomArrayElement(NAMES),
+});
+
+const createComments = () => ({
+  photoComments: Array.from({length: getRandomInteger(1, 6)}, createComment),
+});
+
+const idNumber = createIdGenerator();
+const urlNumber = createIdGenerator();
+const createPhotoDescription = () => ({
+  id: idNumber(),
+  url: `photos/${urlNumber()}.jpg`,
+  description: getRandomArrayElement(DESCRIPTIONS),
+  likes: getRandomInteger(15, 200),
+  comments: createComments(),
+});
+
+const createPhotos = () => ({
+  photoDescriptions: Array.from({length: PHOTO_DESCRIPTION_COUNT}, createPhotoDescription),
+});
+
+export {createPhotos};
